feature/commonjsify-with-umd-export
