+++ .gitignore
@@ -1,6 +1,7 @@
 # Logs
 logs
 *.log
+build
 
 # Runtime data
 pids
