+++ grunt/exec.js
@@ -1,0 +1,6 @@
+'use strict';
+module.exports = {
+    webpack: {
+        cmd: 'webpack'
+    }
+};
