+++ grunt/jscs.js
@@ -1,0 +1,16 @@
+'use strict';
+module.exports = function(grunt) {
+    return {
+        all: {
+            options: {
+                config: '.jscsrc'
+            },
+            files: {
+                src: [
+                    'test/**.js',
+                    'src/**.js',
+                ],
+            }
+        }
+    };
+};
