+++ grunt/jshint.js
@@ -1,0 +1,11 @@
+'use strict';
+module.exports = function(grunt) {
+    return {
+        all: [
+            'src/**/*.js',
+        ],
+        options: {
+            jshintrc: '.jshintrc',
+        }
+    };
+};
