+++ src/routes/index.js
@@ -1,0 +1,9 @@
+'use strict';
+module.exports = {
+    coinstac: {
+        consortia: require('./coinstac/consortia.js')
+    },
+    authentication: require('./authentication.js'),
+    scans: require('./scans.js'),
+    users: require('./users.js'),
+};
